+++ frontend/src/app/router/handles.ts
@@ -77,6 +77,9 @@ export const ENVIRONMENT_V1_ROUTE_DETAIL = `${ENVIRONMENT_V1_ROUTE_DASHBOARD}.de
 
 // --- dashboard/projectV1.ts ---
 export const PROJECT_V1_ROUTE_DETAIL = `${PROJECT_V1_ROUTE_DASHBOARD}.detail`;
+export const PROJECT_V1_ROUTE_INSTANCES = `${PROJECT_V1_ROUTE_DASHBOARD}.instance`;
+export const PROJECT_V1_ROUTE_INSTANCE_CREATE = `${PROJECT_V1_ROUTE_INSTANCES}.create`;
+export const PROJECT_V1_ROUTE_INSTANCE_DETAIL = `${PROJECT_V1_ROUTE_INSTANCES}.detail`;
 export const PROJECT_V1_ROUTE_DATABASES = `${PROJECT_V1_ROUTE_DASHBOARD}.database`;
 export const PROJECT_V1_ROUTE_MASKING_EXEMPTION = `${PROJECT_V1_ROUTE_DASHBOARD}.masking-exemption`;
 export const PROJECT_V1_ROUTE_MASKING_EXEMPTION_CREATE = `${PROJECT_V1_ROUTE_DASHBOARD}.masking-exemption.create`;
