+++ frontend/src/components/instance/DataSourceSection.tsx
@@ -7,7 +7,6 @@ import { useAppStore } from "@/stores/app";
 import { DATASOURCE_READONLY_USER_NAME } from "@/types";
 import { Engine } from "@/types/proto-es/v1/common_pb";
 import { DataSourceType } from "@/types/proto-es/v1/instance_service_pb";
-import { hasWorkspacePermissionV2 } from "@/utils";
 import type { EditDataSource } from "./common";
 import { wrapEditDataSource } from "./common";
 import { DataSourceForm } from "./DataSourceForm";
@@ -36,9 +35,10 @@ export function DataSourceSection({
     editingDataSource,
     readonlyDataSourceList,
     hasReadOnlyDataSource,
+    hasPermission,
   } = ctx;
 
-  const allowUpdate = hasWorkspacePermissionV2("bb.instances.update");
+  const allowUpdate = hasPermission("bb.instances.update");
 
   const handleCreateRODataSource = useCallback(() => {
     if (isCreating) return;
