+++ frontend/src/routes/workspace/InstanceDetailPage.tsx
@@ -1,732 +1,8 @@
-import { create } from "@bufbuild/protobuf";
-import { FieldMaskSchema } from "@bufbuild/protobuf/wkt";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { Trans, useTranslation } from "react-i18next";
-import { useLocation, useNavigate } from "react-router";
-import {
-  AdvancedSearch,
-  getValueFromScopes,
-  type ScopeOption,
-  type SearchParams,
-  type ValueOption,
-} from "@/components/AdvancedSearch";
-import {
-  CreateDatabaseSheet,
-  DatabaseBatchOperationsBar,
-  DatabaseTable,
-  LabelEditorSheet,
-  TransferProjectSheet,
-} from "@/components/database";
-import { EditEnvironmentSheet } from "@/components/EditEnvironmentSheet";
-import { EngineIcon } from "@/components/EngineIcon";
-import { EnvironmentLabel } from "@/components/EnvironmentLabel";
-import { InstanceLabel } from "@/components/InstanceLabel";
-import {
-  InstanceActionDropdown,
-  InstanceFormBody,
-  InstanceFormButtons,
-  InstanceFormProvider,
-  InstanceRoleTable,
-  InstanceSyncButton,
-  useInstanceFormContext,
-} from "@/components/instance";
-import { Alert } from "@/components/ui/alert";
-import { Button } from "@/components/ui/button";
-import { Tabs, TabsList, TabsPanel, TabsTrigger } from "@/components/ui/tabs";
-import { useUnsavedChangesGuard } from "@/hooks/useUnsavedChangesGuard";
-import type { DatabaseFilter } from "@/lib/databaseFilter";
-import {
-  PREPARE_DATABASE_PRODUCT_INTRO,
-  PREPARE_DATABASE_TRANSFER_TIP,
-  PRODUCT_INTRO_TIP_QUERY_KEY,
-  useProductIntro,
-} from "@/lib/productIntro";
-import { pushNotification } from "@/stores";
-import { useAppStore } from "@/stores/app";
-import {
-  environmentNamePrefix,
-  instanceNamePrefix,
-  projectNamePrefix,
-} from "@/stores/modules/v1/common";
-import {
-  isValidDatabaseName,
-  UNKNOWN_ENVIRONMENT_NAME,
-  unknownEnvironment,
-} from "@/types";
-import { State } from "@/types/proto-es/v1/common_pb";
-import type { Database } from "@/types/proto-es/v1/database_service_pb";
-import {
-  BatchUpdateDatabasesRequestSchema,
-  DatabaseSchema$,
-  UpdateDatabaseRequestSchema,
-} from "@/types/proto-es/v1/database_service_pb";
-import { unknownInstance } from "@/types/v1/instance";
-import {
-  extractProjectResourceName,
-  getDefaultPagination,
-  instanceV1Name,
-  setDocumentTitle,
-} from "@/utils";
-
-const instanceHashList = ["overview", "databases", "users"] as const;
-type InstanceHash = (typeof instanceHashList)[number];
-const isInstanceHash = (x: unknown): x is InstanceHash =>
-  instanceHashList.includes(x as InstanceHash);
+import { InstanceDetailView } from "@/components/instance/InstanceDetailView";
+import { instanceNamePrefix } from "@/stores/modules/v1/common";
 
 export function InstanceDetailPage({ instanceId }: { instanceId: string }) {
-  const { t } = useTranslation();
-  const location = useLocation();
-  const navigate = useNavigate();
-  const databasesByName = useAppStore((s) => s.databasesByName);
-  const projectsByName = useAppStore((s) => s.projectsByName);
-  const getDatabaseByName = useAppStore((s) => s.getDatabaseByName);
-  const removeDatabaseMetadataCache = useAppStore(
-    (s) => s.removeDatabaseMetadataCache
-  );
-  const instanceName = `${instanceNamePrefix}${instanceId}`;
-  const cachedInstance = useAppStore((s) => s.instancesByName[instanceName]);
-  const instance = useMemo(
-    () => cachedInstance ?? unknownInstance(),
-    [cachedInstance, instanceName]
-  );
-
-  const routeHash = location.hash.replace(/^#?/, "");
-  const selectedTab: InstanceHash = isInstanceHash(routeHash)
-    ? routeHash
-    : "overview";
-  const [searchParams, setSearchParams] = useState<SearchParams>({
-    query: "",
-    scopes: [{ id: "instance", value: instanceId, readonly: true }],
-  });
-
-  // Selection / batch operations
-  const [selectedNames, setSelectedNames] = useState<Set<string>>(new Set());
-  const [visibleDatabases, setVisibleDatabases] = useState<Database[]>([]);
-  const [refreshToken, setRefreshToken] = useState(0);
-  const autoRefreshCountRef = useRef(0);
-  const [syncing, setSyncing] = useState(false);
-  const [showLabelEditor, setShowLabelEditor] = useState(false);
-  const [showEditEnvDrawer, setShowEditEnvDrawer] = useState(false);
-  const [showTransferDrawer, setShowTransferDrawer] = useState(false);
-  const [showCreateDrawer, setShowCreateDrawer] = useState(false);
-  const [syncingRefreshExhausted, setSyncingRefreshExhausted] = useState(false);
-  const [hasUserProject, setHasUserProject] = useState(false);
-
-  const selectedDatabases = useMemo(() => {
-    if (selectedNames.size === 0) return [];
-    return Array.from(selectedNames)
-      .filter((name) => isValidDatabaseName(name))
-      .map((name) => getDatabaseByName(name));
-  }, [selectedNames, getDatabaseByName, databasesByName]);
-
-  // Mirror `selectedDatabases` into a ref so the batch-operation handlers
-  // below can read the latest value without listing it as a dep. Otherwise
-  // every selection toggle re-creates the handler closures, which cascades
-  // down as fresh prop refs into `DatabaseBatchOperationsBar` and forces
-  // it to re-render (with N selected items, this compounds quickly).
-  const selectedDatabasesRef = useRef(selectedDatabases);
-  selectedDatabasesRef.current = selectedDatabases;
-
-  const refresh = useCallback(() => {
-    setRefreshToken((prev) => prev + 1);
-    setSelectedNames(new Set());
-  }, []);
-
-  const locationQuery = useMemo(
-    () => new URLSearchParams(location.search),
-    [location.search]
-  );
-  const syncingInstanceId = useMemo(
-    () => locationQuery.get("syncingInstance") || undefined,
-    [locationQuery]
-  );
-  const showPrepareDatabaseTip =
-    locationQuery.get(PRODUCT_INTRO_TIP_QUERY_KEY) ===
-    PREPARE_DATABASE_TRANSFER_TIP;
-
-  useEffect(() => {
-    if (!syncingInstanceId || visibleDatabases.length > 0) {
-      setSyncingRefreshExhausted(false);
-      return;
-    }
-    setSyncingRefreshExhausted(false);
-    autoRefreshCountRef.current = 0;
-    const timer = window.setInterval(() => {
-      autoRefreshCountRef.current += 1;
-      refresh();
-      if (autoRefreshCountRef.current >= 12) {
-        setSyncingRefreshExhausted(true);
-        window.clearInterval(timer);
-      }
-    }, 5000);
-    return () => window.clearInterval(timer);
-  }, [syncingInstanceId, visibleDatabases.length, refresh]);
-
-  const handleSyncSchema = useCallback(async () => {
-    if (syncing) return;
-    setSyncing(true);
-    pushNotification({
-      module: "bytebase",
-      style: "INFO",
-      title: t("db.start-to-sync-schema"),
-    });
-    try {
-      await useAppStore
-        .getState()
-        .batchSyncDatabases(Array.from(selectedNames));
-      for (const name of selectedNames) {
-        removeDatabaseMetadataCache(name);
-      }
-      pushNotification({
-        module: "bytebase",
-        style: "SUCCESS",
-        title: t("db.successfully-synced-schema"),
-      });
-      setSelectedNames(new Set());
-    } catch {
-      pushNotification({
-        module: "bytebase",
-        style: "CRITICAL",
-        title: t("db.failed-to-sync-schema"),
-      });
-    } finally {
-      setSyncing(false);
-    }
-  }, [syncing, selectedNames, removeDatabaseMetadataCache, t]);
-
-  const handleLabelsApply = useCallback(
-    async (labelsList: { [key: string]: string }[]) => {
-      try {
-        await useAppStore.getState().batchUpdateDatabases(
-          create(BatchUpdateDatabasesRequestSchema, {
-            parent: "-",
-            requests: selectedDatabasesRef.current.map((database, i) =>
-              create(UpdateDatabaseRequestSchema, {
-                database: create(DatabaseSchema$, {
-                  ...database,
-                  labels: labelsList[i],
-                }),
-                updateMask: create(FieldMaskSchema, { paths: ["labels"] }),
-              })
-            ),
-          })
-        );
-        refresh();
-        pushNotification({
-          module: "bytebase",
-          style: "SUCCESS",
-          title: t("common.updated"),
-        });
-      } catch {
-        pushNotification({
-          module: "bytebase",
-          style: "CRITICAL",
-          title: t("common.failed"),
-        });
-      }
-    },
-    [refresh, t]
-  );
-
-  const handleEnvironmentUpdate = useCallback(
-    async (environment: string) => {
-      try {
-        await useAppStore.getState().batchUpdateDatabases(
-          create(BatchUpdateDatabasesRequestSchema, {
-            parent: "-",
-            requests: selectedDatabasesRef.current.map((database) =>
-              create(UpdateDatabaseRequestSchema, {
-                database: create(DatabaseSchema$, {
-                  name: database.name,
-                  environment,
-                }),
-                updateMask: create(FieldMaskSchema, { paths: ["environment"] }),
-              })
-            ),
-          })
-        );
-        refresh();
-        pushNotification({
-          module: "bytebase",
-          style: "SUCCESS",
-          title: t("common.updated"),
-        });
-      } catch {
-        pushNotification({
-          module: "bytebase",
-          style: "CRITICAL",
-          title: t("common.failed"),
-        });
-      }
-    },
-    [refresh, t]
-  );
-
-  const handleTransferProject = useCallback(
-    async (projectName: string) => {
-      try {
-        await useAppStore.getState().batchUpdateDatabases(
-          create(BatchUpdateDatabasesRequestSchema, {
-            parent: "-",
-            requests: selectedDatabasesRef.current.map((database) =>
-              create(UpdateDatabaseRequestSchema, {
-                database: create(DatabaseSchema$, {
-                  name: database.name,
-                  project: projectName,
-                }),
-                updateMask: create(FieldMaskSchema, { paths: ["project"] }),
-              })
-            ),
-          })
-        );
-        refresh();
-        pushNotification({
-          module: "bytebase",
-          style: "SUCCESS",
-          title: t("database.successfully-transferred-databases"),
-        });
-      } catch {
-        pushNotification({
-          module: "bytebase",
-          style: "CRITICAL",
-          title: t("common.failed"),
-        });
-      }
-    },
-    [refresh, t]
-  );
-  // Trigger a fetch on mount so the instance is hydrated into the
-  // `useAppStore` cache. Without this, hard-refreshing the page shows
-  // "Unknown instance" because the cache hasn't been populated yet.
-  useEffect(() => {
-    void useAppStore.getState().getOrFetchInstanceByName(instanceName);
-  }, [instanceName]);
-
-  // Canonicalize legacy query state and invalid/missing hashes. The location
-  // remains the tab source of truth, so Back/Forward updates the selected tab.
-  useEffect(() => {
-    const query = new URLSearchParams(location.search);
-    query.delete("qs");
-    const search = query.toString();
-    const nextSearch = search ? `?${search}` : "";
-    const nextHash = `#${selectedTab}`;
-    if (location.search === nextSearch && location.hash === nextHash) return;
-    void navigate(
-      {
-        pathname: location.pathname,
-        search: nextSearch,
-        hash: nextHash,
-      },
-      { replace: true }
-    );
-  }, [
-    location.hash,
-    location.pathname,
-    location.search,
-    navigate,
-    selectedTab,
-  ]);
-
-  // Set document title
-  useEffect(() => {
-    if (instance.title) {
-      setDocumentTitle(instance.title);
-    }
-  }, [instance.title]);
-
-  const syncSchema = useCallback(
-    async (enableFullSync: boolean) => {
-      await useAppStore.getState().syncInstance(instance.name, enableFullSync);
-      useAppStore.getState().removeCacheByInstance(instance.name);
-    },
-    [instance.name]
-  );
-
-  // Database filter
-  const envVal = getValueFromScopes(searchParams, "environment");
-  const selectedEnvironment = envVal
-    ? `${environmentNamePrefix}${envVal}`
-    : undefined;
-  const projectVal = getValueFromScopes(searchParams, "project");
-  const selectedProject = projectVal
-    ? `${projectNamePrefix}${projectVal}`
-    : undefined;
-  const selectedLabels = useMemo(
-    () =>
-      searchParams.scopes.filter((s) => s.id === "label").map((s) => s.value),
-    [searchParams]
-  );
-
-  const filter: DatabaseFilter = useMemo(
-    () => ({
-      environment: selectedEnvironment,
-      project: selectedProject,
-      query: searchParams.query,
-      labels: selectedLabels.length > 0 ? selectedLabels : undefined,
-    }),
-    [selectedEnvironment, selectedProject, searchParams.query, selectedLabels]
-  );
-
-  const environments = useAppStore((s) => s.environmentList ?? []);
-
-  // Reactive: the actuator's `defaultProject` is fetched asynchronously, so
-  // we subscribe through the store selector — otherwise the value is
-  // captured as `""` on first render and the API filter becomes broken.
-  const defaultProjectId = useAppStore((s) =>
-    extractProjectResourceName(s.serverInfo?.defaultProject ?? "")
-  );
-  const defaultProjectName = useAppStore(
-    (s) => s.serverInfo?.defaultProject ?? ""
-  );
-  const unassignedProjectOption = useMemo<ValueOption>(
-    () => ({
-      value: defaultProjectId,
-      keywords: ["unassigned", "default"],
-      custom: true,
-      render: () => (
-        <span className="italic text-control-light">
-          {t("common.unassigned")}
-        </span>
-      ),
-    }),
-    [defaultProjectId, t]
-  );
-  const searchProjects = useCallback(
-    async (keyword: string): Promise<ValueOption[]> => {
-      const { projects } = await useAppStore.getState().fetchProjectList({
-        pageSize: getDefaultPagination(),
-        filter: keyword.trim() ? { query: keyword } : undefined,
-      });
-      return projects.map<ValueOption>((p) => {
-        const id = extractProjectResourceName(p.name);
-        if (id === defaultProjectId) return unassignedProjectOption;
-        return {
-          value: id,
-          keywords: [id, p.title],
-        };
-      });
-    },
-    [defaultProjectId, unassignedProjectOption]
-  );
-
-  const scopeOptions: ScopeOption[] = useMemo(
-    () => [
-      {
-        id: "environment",
-        title: t("common.environment"),
-        description: t("issue.advanced-search.scope.environment.description"),
-        options: [unknownEnvironment(), ...environments].map((env) => {
-          const isUnknown = env.name === UNKNOWN_ENVIRONMENT_NAME;
-          return {
-            value: env.id,
-            keywords: isUnknown
-              ? ["unassigned", "none", env.id]
-              : [env.id, env.title],
-            custom: true,
-            render: () => <EnvironmentLabel environment={env} />,
-          };
-        }),
-      },
-      {
-        id: "project",
-        title: t("common.project"),
-        description: t("issue.advanced-search.scope.project.description"),
-        // Static option lets the selected-tag display resolve the default
-        // project id to "Unassigned".
-        options: [unassignedProjectOption],
-        onSearch: searchProjects,
-      },
-      {
-        id: "label",
-        title: t("common.labels"),
-        description: t("issue.advanced-search.scope.label.description"),
-        allowMultiple: true,
-      },
-    ],
-    [t, environments, searchProjects, unassignedProjectOption]
-  );
-
-  const handleTabChange = useCallback(
-    (tab: string | number | null) => {
-      if (typeof tab !== "string" || !isInstanceHash(tab)) return;
-      void navigate(
-        {
-          pathname: location.pathname,
-          search: location.search,
-          hash: `#${tab}`,
-        },
-        { replace: true }
-      );
-    },
-    [location.pathname, location.search, navigate]
-  );
-
-  useEffect(() => {
-    if (!syncingInstanceId || visibleDatabases.length === 0) {
-      setHasUserProject(false);
-      return;
-    }
-
-    const isUserProject = (project?: { name?: string }) =>
-      !!defaultProjectName &&
-      !!project?.name &&
-      project.name !== defaultProjectName;
-
-    if (Object.values(projectsByName).some(isUserProject)) {
-      setHasUserProject(true);
-      return;
-    }
-
-    let cancelled = false;
-    void (async () => {
-      try {
-        const { projects } = await useAppStore.getState().fetchProjectList({
-          pageSize: getDefaultPagination(),
-          silent: true,
-        });
-        if (cancelled) {
-          return;
-        }
-        setHasUserProject(projects.some(isUserProject));
-      } catch {
-        if (!cancelled) {
-          setHasUserProject(false);
-        }
-      }
-    })();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [
-    defaultProjectName,
-    projectsByName,
-    syncingInstanceId,
-    visibleDatabases.length,
-  ]);
-
-  const showSyncingInstanceHint =
-    !!syncingInstanceId &&
-    visibleDatabases.length === 0 &&
-    !syncingRefreshExhausted;
-  const showNoDatabasesAfterSync =
-    !!syncingInstanceId &&
-    visibleDatabases.length === 0 &&
-    syncingRefreshExhausted;
-  const showPostSyncTransferAction =
-    !!syncingInstanceId && visibleDatabases.length > 0 && hasUserProject;
-  const shouldShowTransferIntro =
-    showPrepareDatabaseTip && showPostSyncTransferAction;
-  useProductIntro({
-    id: PREPARE_DATABASE_PRODUCT_INTRO,
-    title: t("workspace-setup-guide.intro.transfer-title"),
-    description: t("workspace-setup-guide.intro.transfer-description"),
-    disabled: !shouldShowTransferIntro,
-  });
-  const syncingInstanceName = syncingInstanceId
-    ? `${instanceNamePrefix}${syncingInstanceId}`
-    : undefined;
-  const syncingInstance =
-    syncingInstanceId === instanceId ? instance : undefined;
-  const handleTransferSyncedDatabases = useCallback(() => {
-    setSelectedNames(
-      new Set(visibleDatabases.map((database) => database.name))
-    );
-    setShowTransferDrawer(true);
-  }, [visibleDatabases]);
-
   return (
-    <div className="p-4 flex flex-col gap-y-2">
-      {/* Archive banner */}
-      {instance.state === State.DELETED && (
-        <div className="bg-gray-700 text-white text-center py-2 rounded-sm text-sm font-medium">
-          {t("common.archived")}
-        </div>
-      )}
-
-      {/* No environment warning */}
-      {!instance.environment && (
-        <Alert
-          variant="warning"
-          className="mb-4"
-          description={t("instance.no-environment")}
-        />
-      )}
-
-      {/* Header */}
-      <div className="flex items-center justify-between">
-        <div className="flex items-center gap-x-2">
-          <EngineIcon engine={instance.engine} className="h-6 w-6" />
-          <span className="text-lg font-medium">
-            {instanceV1Name(instance)}
-          </span>
-        </div>
-        <div className="flex items-center gap-x-2">
-          {instance.state === State.ACTIVE && (
-            <InstanceSyncButton
-              instanceName={instance.name}
-              instanceTitle={instance.title}
-              onSyncSchema={syncSchema}
-            />
-          )}
-          <InstanceActionDropdown instance={instance} />
-        </div>
-      </div>
-
-      {/* Tabs */}
-      <Tabs value={selectedTab} onValueChange={handleTabChange}>
-        <TabsList className="border-b-0">
-          <TabsTrigger value="overview">{t("common.overview")}</TabsTrigger>
-          <TabsTrigger value="databases">{t("common.databases")}</TabsTrigger>
-          <TabsTrigger value="users">{t("instance.users")}</TabsTrigger>
-        </TabsList>
-
-        <TabsPanel value="overview">
-          <InstanceFormProvider instance={instance}>
-            <InstanceFormBody />
-            <InstanceFormButtons />
-            <UnsavedChangesGuard />
-          </InstanceFormProvider>
-        </TabsPanel>
-
-        <TabsPanel value="databases" keepMounted={false}>
-          <div className="flex flex-col gap-y-2">
-            {showSyncingInstanceHint && (
-              <Alert
-                variant="info"
-                title={
-                  <Trans
-                    t={t}
-                    i18nKey="db.project-instance-syncing-title"
-                    components={{
-                      instance: (
-                        <InstanceLabel
-                          instance={syncingInstance}
-                          instanceName={syncingInstanceName}
-                        />
-                      ),
-                    }}
-                  />
-                }
-                description={
-                  <div className="flex flex-col gap-y-3 sm:flex-row sm:items-center sm:justify-between sm:gap-x-4">
-                    <span>{t("db.project-instance-syncing-description")}</span>
-                    <Button size="sm" appearance="outline" onClick={refresh}>
-                      {t("common.refresh")}
-                    </Button>
-                  </div>
-                }
-              />
-            )}
-            {showPostSyncTransferAction && (
-              <Alert
-                variant="info"
-                title={t("db.instance-databases-synced-title")}
-                description={
-                  <div className="flex flex-col gap-y-3 sm:flex-row sm:items-center sm:justify-between sm:gap-x-4">
-                    <span>{t("db.instance-databases-synced-description")}</span>
-                    <Button size="sm" onClick={handleTransferSyncedDatabases}>
-                      {t("db.instance-databases-synced-action")}
-                    </Button>
-                  </div>
-                }
-              />
-            )}
-            {showNoDatabasesAfterSync && (
-              <Alert
-                variant="info"
-                title={t("db.instance-no-databases-title")}
-                description={
-                  <div className="flex flex-col gap-y-3 sm:flex-row sm:items-center sm:justify-between sm:gap-x-4">
-                    <span>{t("db.instance-no-databases-description")}</span>
-                    <Button
-                      size="sm"
-                      appearance="outline"
-                      onClick={() => setShowCreateDrawer(true)}
-                    >
-                      {t("db.instance-no-databases-action")}
-                    </Button>
-                  </div>
-                }
-              />
-            )}
-            <AdvancedSearch
-              params={searchParams}
-              onParamsChange={setSearchParams}
-              placeholder={t("database.filter-database")}
-              scopeOptions={scopeOptions}
-            />
-            <EditEnvironmentSheet
-              open={showEditEnvDrawer}
-              onClose={() => setShowEditEnvDrawer(false)}
-              onUpdate={handleEnvironmentUpdate}
-            />
-            <LabelEditorSheet
-              open={showLabelEditor}
-              databases={selectedDatabases}
-              onClose={() => setShowLabelEditor(false)}
-              onApply={handleLabelsApply}
-            />
-            <TransferProjectSheet
-              open={showTransferDrawer}
-              databases={selectedDatabases}
-              onClose={() => setShowTransferDrawer(false)}
-              onTransfer={handleTransferProject}
-            />
-            <CreateDatabaseSheet
-              open={showCreateDrawer}
-              onClose={() => setShowCreateDrawer(false)}
-              instanceName={instanceName}
-            />
-            <DatabaseTable
-              filter={filter}
-              parent={instance.name}
-              mode="ALL"
-              selectedNames={selectedNames}
-              onSelectedNamesChange={setSelectedNames}
-              onDatabasesChange={setVisibleDatabases}
-              refreshToken={refreshToken}
-              selectionColumnIntroTarget={
-                shouldShowTransferIntro
-                  ? PREPARE_DATABASE_PRODUCT_INTRO
-                  : undefined
-              }
-            />
-            {/* Batch operations bar is fixed within the visible main content. */}
-            <DatabaseBatchOperationsBar
-              databases={selectedDatabases}
-              onSyncSchema={handleSyncSchema}
-              onEditLabels={() => setShowLabelEditor(true)}
-              onEditEnvironment={() => setShowEditEnvDrawer(true)}
-              onTransferProject={() => setShowTransferDrawer(true)}
-              allSelected={
-                visibleDatabases.length > 0 &&
-                visibleDatabases.every((d) => selectedNames.has(d.name))
-              }
-              onToggleSelectAll={() => {
-                const allOnPage =
-                  visibleDatabases.length > 0 &&
-                  visibleDatabases.every((d) => selectedNames.has(d.name));
-                if (allOnPage) setSelectedNames(new Set());
-                else
-                  setSelectedNames(
-                    new Set(visibleDatabases.map((d) => d.name))
-                  );
-              }}
-            />
-          </div>
-        </TabsPanel>
-
-        <TabsPanel value="users">
-          <InstanceRoleTable instanceRoleList={instance.roles ?? []} />
-        </TabsPanel>
-      </Tabs>
-    </div>
+    <InstanceDetailView instanceName={`${instanceNamePrefix}${instanceId}`} />
   );
-}
-
-function UnsavedChangesGuard() {
-  const { valueChanged } = useInstanceFormContext();
-  useUnsavedChangesGuard(valueChanged);
-  return null;
 }
